pyproject.toml
[tool.poetry]
name = "message_encryptor"
version = "0.1.0"
description = ""
authors = ["Daniel Neshyba-Rowe <rowedanielp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
